fix: cambios menores en files y words

+++ helpers/deleteFiles.js
@@ -5,11 +5,12 @@ const path = require('path')
  * @param {string} url url del archivo a borrar 
  */
 const deleteFiles = (url) => {
-   const filePath = url.replace(process.env.URL_BASE, '');
-   const rootPath = path.join('public', filePath);
-
    try {
-      fs.unlinkSync(rootPath);
+      const filePath = url.replace(process.env.URL_BASE, '');
+      const rootPath = path.join('public', filePath);
+      if (fs.existsSync(rootPath)) {
+         fs.unlinkSync(rootPath);
+      }
    } catch (error) {
       console.log(error);
       throw error;
